Initial Check in of application
Two page application one for students, one for an admin.

The index.html page is for student viewing of events.  The admin page
uses the docs service for CRUD operations on 'events'.

Currently uses the auth service to anonymously retrieve a token on the
index.html page for event viewing.  Currently uses a hard coded admin
user to retrieve a token using the auth service in the admin.html page.